added unit test case
